remove unused code and views
