+++ app/assets/javascripts/application.js
@@ -18,7 +18,6 @@
 //= require bp_ontolobridge
 //= require bp_form_complete
 //= require bp_analytics
-//= require bp_user_intention_survey
 //= require bp_search
 //= require bp_mappings
 //= require bp_admin
