+++ app/controllers/admin_controller.rb
@@ -3,17 +3,6 @@ class AdminController < ApplicationController
   layout :determine_layout
   before_action :cache_setup
 
-  DEBUG_BLACKLIST = [:"$,", :$ADDITIONAL_ONTOLOGY_DETAILS, :$rdebug_state,
-                     :$PROGRAM_NAME, :$LOADED_FEATURES, :$KCODE, :$-i, :$rails_rake_task, :$$, :$gems_build_rake_task,
-                     :$daemons_stop_proc, :$VERBOSE, :$DAEMONS_ARGV, :$daemons_sigterm,
-                     :$DEBUG_BEFORE, :$stdout, :$-0, :$-l, :$-I, :$DEBUG, :$',
-                     :$gems_rake_task, :$_, :$CODERAY_DEBUG, :$-F, :$", :$0, :$=, :$FILENAME, :$?,
-                     :$!, :$rdebug_in_irb, :$-K, :$TESTING, :$fileutils_rb_have_lchmod,
-                     :$binding, :$-v, :$>, :$SAFE, :$/,
-                     :$fileutils_rb_have_lchown, :$-p, :$-W, :$:, :$__dbg_interface,
-                     :$stderr, :$\, :$&, :$<, :$debug, :$;, :$~, :$-a,
-                     :$DEBUG_RDOC, :$CGI_ENV, :$LOAD_PATH, :$-d, :$*, :$., :$-w, :$+,
-                     :$@, :$`, :$stdin, :$1, :$2, :$3, :$4, :$5, :$6, :$7, :$8, :$9]
   ADMIN_URL = "#{LinkedData::Client.settings.rest_url}/admin/"
   ONTOLOGIES_URL = "#{ADMIN_URL}ontologies_report"
   USERS_URL = "#{LinkedData::Client.settings.rest_url}/users"
