+++ app/controllers/ajax_proxy_controller.rb
@@ -64,9 +64,6 @@ class AjaxProxyController < ApplicationController
     render_json simple_ontology.to_json
   end
 
-  def loading_spinner
-    render :partial => "loading_spinner"
-  end
 
   private
