+++ app/controllers/concepts_controller.rb
@@ -175,8 +175,6 @@ class ConceptsController < ApplicationController
     @concept = @ontology.explore.single_class({full: true}, params[:conceptid])
     concept_not_found(params[:conceptid]) if @concept.nil?
 
-    @immediate_load = true
-
     render partial: "biomixer", layout: false
   end
 
@@ -185,9 +183,6 @@ private
 
   def show_ajax_request
     case params[:callback]
-    when 'load' # Load pulls in all the details of a node
-      gather_details
-      render :partial => 'load'
     when 'children' # Children is called only for drawing the tree
       @children = @concept.explore.children(pagesize: 750, concept_schemes: Array(@schemes).join(','), language: request_lang, display: 'prefLabel,obsolete,hasChildren').collection || []
       @children.sort! { |x, y| (x.prefLabel || "").downcase <=> (y.prefLabel || "").downcase } unless @children.empty?
