+++ app/controllers/home_controller.rb
@@ -39,10 +39,6 @@ class HomeController < ApplicationController
 
   end
 
-  def render_layout_partial
-    partial = params[:partial]
-    render partial: "layouts/#{partial}"
-  end
 
   def all_resources
     @conceptid = params[:conceptid]
@@ -109,9 +105,6 @@ class HomeController < ApplicationController
     end
   end
 
-  def user_intention_survey
-    render partial: 'user_intention_survey', layout: false
-  end
 
   def site_config
     render json: bp_config_json
@@ -139,13 +132,6 @@ class HomeController < ApplicationController
   end
 
   def feedback_complete; end
-
-  def validate_ontology_file_show; end
-
-  def validate_ontology_file
-    response = LinkedData::Client::HTTP.post('/validate_ontology_file', ontology_file: params[:ontology_file])
-    @process_id = response.process_id
-  end
 
   def annotator_recommender_form
     if params[:submit_button] == "annotator"
