--- app/models/analytics.rb
@@ -1,3 +1,0 @@
-class Analytics < ApplicationRecord
-  serialize :params
-end
