--- app/views/admin/index.html.erb.old
@@ -1,84 +1,0 @@
-<script type="text/javascript">
-    function resetMemcacheConnection(){
-        ajaxCall("resetcache", "#reset_memcache_status");
-    }
-
-    function flushMemcache(){
-        ajaxCall("clearcache", "#flush_memcache_status");
-    }
-
-    function ajaxCall(operation, status){
-        $(status).html("");
-        $.ajax({
-            type: "POST",
-            url: "/admin/" + operation,
-            data: "",
-            dataType: "text",
-            beforeSend: function(XMLHttpRequest){
-                updateStatus(XMLHttpRequest, status)
-            },
-            success: function(data, msg){
-                $(status).text(data).html();
-            },
-            error: function(XMLHttpRequest, textStatus, errorThrown){
-                $(status).html(errorThrown + " " + textStatus);
-            }
-        });
-    };
-
-    function updateStatus(XMLHttpRequest, status){
-        $(status).html("Loading");
-    }
-</script>
-<style>
-    .admin_utilities td {
-        padding: .5em;
-    }
-</style>
-
-
-
-
-<div class="feedback" style="width: 80%; padding: 2em;">
-    <%= render :partial => 'memcache_stats' if @cache.respond_to?(:stats)%>
-    <div>
-        <h2>Memcache Utilities</h2>
-        <table class="admin_utilities" style="float: left;">
-            <tr>
-                <td>
-                    <input type="button" onclick="flushMemcache();" value="Flush Memcache">
-                </td>
-                <td>
-                    <span id="flush_memcache_status"></span>
-                </td>
-            </tr>
-            <tr>
-                <td>
-                    <input type="button" onclick="resetMemcacheConnection();" value="Reset Memcache Connection">
-                </td>
-                <td>
-                    <span id="reset_memcache_status"></span>
-                </td>
-            </tr>
-        </table>
-    </div>
-</div>
-
-
-
-
-
-
-
-<div id="container" style="clear:left">
-  <h2>Ontology Administration</h2>
-
-</div>
-
-
-<div id="global_debug" style="clear: both; padding: 2em;">
-  <h2>Global Variables / Settings</h2>
-  <pre>
-    <%=PP.pp(@globals, "")%>
-  </pre>
-</div>
