--- app/views/ajax_proxy/_loading_spinner.html.haml
@@ -1,1 +1,0 @@
-=loading_spinner()
