+++ app/views/concepts/_biomixer.html.erb
@@ -2,7 +2,7 @@
 <% user_api_key = (session[:user].try(:apikey) || $BIOMIXER_APIKEY).to_s %>
 <% rest_domain = $REST_URL.sub(/https?:\/\//, "") %>
 <% src_url = "#{$BIOMIXER_URL}/?mode=embed&embed_mode=paths_to_root&ontology_acronym=#{@ontology.acronym}&full_concept_id=#{CGI.escape(@concept.fullId)}&userapikey=#{user_api_key}&restURLPrefix=#{rest_domain}" %>
-<% original_src = @immediate_load ? src_url : "/ajax/loading_spinner"%>
+<% original_src = src_url%>
 
 <script>
 jQuery(document).data().bp.biomixer_fullscreen = {};
