--- app/views/concepts/_flexviz.html.erb
@@ -1,104 +1,0 @@
-<% params['height'] ||= "500"%>
-<% params['width'] ||= "500" %>
-
-<!--  BEGIN Browser History required section -->
-<style type="text/css">
-  #ie_historyFrame { width: 0px; height: 0px; display:none }
-  #firefox_anchorDiv { width: 0px; height: 0px; display:none }
-  #safari_formDiv { width: 0px; height: 0px; display:none }
-  #safari_rememberDiv { width: 0px; height: 0px; display:none }
-</style>
-<script src="<%=$FLEX_URL%>/history/history.js" language="javascript"></script>
-<!--  END Browser History required section -->
-
-<script language="JavaScript" type="text/javascript">
-  <!--
-  // Globals
-  // Major version of Flash required
-  var requiredMajorVersion = 9;
-  // Minor version of Flash required
-  var requiredMinorVersion = 0;
-  // Minor version of Flash required
-  var requiredRevision = 28;
-  // -->
-</script>
-
-<script src="<%=$FLEX_URL%>/AC_OETags.js" language="javascript" type="text/javascript"></script>
-
-<script language="javascript" type="text/javascript">
-// <!--
-// Version check for the Flash Player that has the ability to start Player Product Install (6.0r65)
-var hasProductInstall = DetectFlashVer(6, 0, 65);
-
-// Version check based upon the values defined in globals
-var hasRequestedVersion = DetectFlashVer(requiredMajorVersion, requiredMinorVersion, requiredRevision);
-
-if ( hasProductInstall && !hasRequestedVersion ) {
-	// DO NOT MODIFY THE FOLLOWING FOUR LINES
-	// Location visited after installation is complete if installation is required
-	var MMPlayerType = (isIE == true) ? "ActiveX" : "PlugIn";
-	var MMredirectURL = window.location;
-    document.title = document.title.slice(0, 47) + " - Flash Player Installation";
-    var MMdoctitle = document.title;
-
-	AC_FL_RunContent(
-		"src", "playerProductInstall",
-		"FlashVars", "MMredirectURL="+MMredirectURL+'&MMplayerType='+MMPlayerType+'&MMdoctitle='+MMdoctitle+"",
-    "width", "<%=params['height']%>px",
-		"height", "<%=params['width']%>px",
-		"align", "left",
-		"id", "FlexoViz",
-		"quality", "high",
-		"bgcolor", "#ffffff",
-		"name", "FlexoViz",
-		"allowScriptAccess","sameDomain",
-		"type", "application/x-shockwave-flash",
-		"pluginspage", "http://www.adobe.com/go/getflashplayer"
-	);
-} else if (hasRequestedVersion) {
-	// if we've detected an acceptable version embed the Flash Content SWF when all tests are passed
-	AC_FL_RunContent(
-      "src", "<%=$FLEX_URL%>/BasicFlexoViz?v=2.3.4.1",
-			"width", "<%=params['height']%>px",
-			"height", "<%=params['width']%>px",
-			"align", "left",
-			"id", "FlexoViz",
-			"quality", "high",
-			"bgcolor", "#ffffff",
-			"name", "FlexoViz",
-			"allowScriptAccess","always",
-			"type", "application/x-shockwave-flash",
-			"flashVars", "virtual=true&server=<%=$LEGACY_REST_URL%>&ontology=<%=params['ontologyid']%>&nodeid=<%=params['conceptid']%>",
-			"pluginspage", "http://www.adobe.com/go/getflashplayer"
-	);
-  } else {
-  	// flash is too old or we can't detect the plugin
-    var alternateContent = '<br/>This website requires the Adobe Flash Player.  '
-    + 'Please download and install the Flash plug-in and try again.  '
-   	+ '<a href=http://www.adobe.com/go/getflash/>Get Flash</a><br/>';
-    document.write(alternateContent);
-  }
-// -->
-</script>
-<noscript>
-  	<object classid="clsid:D27CDB6E-AE6D-11cf-96B8-444553540000"
-			id="FlexoViz" width="<%=params['width']%>px" height="<%=params['height']%>px"
-			codebase="http://fpdownload.macromedia.com/get/flashplayer/current/swflash.cab">
-      <param name="movie" value="<%=$FLEX_URL%>/BasicFlexoViz.swf" />
-			<param name="quality" value="high" />
-			<param name="bgcolor" value="#ffffff" />
-			<param name="allowScriptAccess" value="always" />
-      <param name="flashVars" value="server=<%=$LEGACY_REST_URL%>&ontology=<%=params['ontologyid']%>&nodeid=<%=params['conceptid']%>" />
-			<embed src="FlexoViz.swf" bgcolor="#ffffff"
-				width="<%=params['width']%>px" height="<%=params['height']%>px" name="FlexoViz" align="left"
-				play="true"
-				loop="false"
-				quality="high"
-				allowScriptAccess="always"
-				type="application/x-shockwave-flash"
-				flashVars="server=&ontology=&nodeid="
-				pluginspage="http://www.adobe.com/go/getflashplayer">
-			</embed>
-	</object>
-</noscript>
-
