--- app/views/concepts/_images.html.erb
@@ -1,21 +1,0 @@
-
-<object classid="clsid:D27CDB6E-AE6D-11cf-96B8-444553540000"
-		id="BasicFlexoViz" width="100%" height="450"
-		codebase="http://fpdownload.macromedia.com/get/flashplayer/current/swflash.cab">
-		<param name="movie" value="<%=$FLEX_URL%>/BasicFlexoViz.swf?v=<%=$FLEX_VERSION%>" />
-		<param name="quality" value="high" />
-		<param name="bgcolor" value="#ffffff" />
-		<param name="allowScriptAccess" value="always" />
-		<param name=FlashVars value="ontology=<%=@ontology.id%>&nodeid=<%=@concept.id%>&server=<%=$REST_DOMAIN%>&apikey=<%=$FLEXOVIZ_APIKEY%>" />
-		<embed src="<%=$FLEX_URL%>/BasicFlexoViz.swf?v=<%=$FLEX_VERSION%>" bgcolor="#ffffff"
-			width="100%" height="450" name="BasicFlexoViz" align="middle"
-			play="true"
-			loop="false"
-			quality="high"
-			allowScriptAccess="always"
-			type="application/x-shockwave-flash"
-			flashVars="ontology=<%=@ontology.id%>&nodeid=<%=@concept.id%>&server=<%=$REST_DOMAIN%>&apikey=<%=$FLEXOVIZ_APIKEY%>"
-			pluginspage="http://www.adobe.com/go/getflashplayer">
-		</embed>
-</object>
-
