--- app/views/concepts/_load.html.haml
@@ -1,21 +1,0 @@
-- require 'cgi'
-- #Created a different way to parse because of the limitations of JSon to do large HTML segments ||| divides the areas , final 2 are mappings and notes to be updated
-
-#{render :partial =>'/concepts/biomixer'}
-|||
-#{render :partial =>'details'}
-|||
-#{render partial: '/notes/list'}
-|||
-#{render :partial =>'/mappings/concept_mappings'}
-|||
-#{render :partial =>'/concepts/request_term'}
-|||
-/ # This used to hold demo-specific info, now it's a placeholder
-&nbsp;
-|||
-/ # default notes count = 0
-=0
-|||
--# TODO_REV support mappings
--# #{@concept.mapping_count}
