--- app/views/concepts/_software.html.erb
@@ -1,23 +1,0 @@
-<table class="metadata" width="100%" cellpadding="0" cellspacing="0">
-	
-	<%for soft in @software%>
-	<tr>
-		<td colspan="2"style="border-bottom:1px solid #9F9F9F;font-size:17px;font-weight:bold;"><a href="<%=soft.url%>" style="text-decoration:none;"><%= soft.name%></td>
-	</tr>
-	<tr style="background-color:#EFEFEF">
-		<td  colspan="2" style="padding:10px;border-bottom:1px solid #9F9F9F"><%=soft.description%></td>
-	</tr>
-	<tr>
-		<th style="width:100px;text-align:left;padding-top:5px;">Authors:</th><td style="padding-top:5px;"><%=soft.authors%></td>
-	</tr>
-	<tr>
-		<th style="width:100px;text-align:left;padding-top:5px;">Organization:</th><td style="padding-top:5px;"><%=soft.organization%></td>
-	</tr>
-	<tr>
-		<th style="width:100px;text-align:left;padding-top:5px;">License:</th><td style="padding-top:5px;"><%=soft.license%></td>
-	</tr>
-	<tr>
-		<td colspan="2" style="width:100%;border-bottom:2px solid #9F9F9F;margin-bottom:15px;">&nbsp;</td>
-	</tr>
-	<%end%>
-</table>
