--- app/views/home/_user_intention_survey.html.haml
@@ -1,37 +1,0 @@
-- case $USER_INTENTION_SURVEY
-- when :accomplish
-  %h2{style: "margin-top: -1em; font-size: 32px; color: #338D0C;"} Help Us Make BioPortal Better
-  %p{style: "font-size: larger;"}
-    We are collecting information on how our community uses BioPortal.
-    To help us improve BioPortal, please answer the following question.
-  %form
-    %p
-      %h2{style: "margin-bottom: 0;"} What are you trying to accomplish with BioPortal during this visit?
-      %textarea{id: "intention_response", style: "width: 600px; height: 225px; padding: 3px;"}
-    %p
-      %input{type: "checkbox", name: "dont_show_again", id: "dont_show_again"}
-      %label{for: "dont_show_again"} Don't show me this dialog again
-    %p
-      %input{id: "intention_submit", type: "button", value: "Submit"}
-      %input{id: "intention_close", type: "button", value: "Close"}
-- when :external_july_2014
-  %h2{style: "font-weight: 700; margin-top: -1em; font-size: 32px; color: green;"} User Survey Summer 2014
-  %div.user_intention_survey_text
-    %p Please help us evaluate the utility and usability of NCBO technologies by taking a short (10 minute) survey.
-    %p The survey can be anonymous but if you choose to provide your email address, you can be entered into a prize drawing.
-    %form{style: "margin-top: 2em;"}
-      %input{id: "intention_submit", type: "button", value: "Take Survey Now", class: "link_button", style: "border: medium solid; font-size: larger;"}
-      %br
-      -# %input{id: "intention_remind", type: "button", value: "Remind Me Later", class: "link_button"}
-      -# %input{id: "intention_close", type: "button", value: "Close", class: "link_button"}
-      %br
-      %br
-      %div{style: "font-size: 13px;"}
-        %input{type: "radio", name: "popup_later", value: "remind_me", id: "remind_me", checked: true}
-        %label{for: "remind_me"} Remind me later
-        %input{type: "radio", name: "popup_later", value: "dont_show_again", id: "dont_show_again"}
-        %label{for: "dont_show_again"} Don't ask again
-
-  :javascript
-    jQuery(".user_intention_survey_text .link_button").button()
-
