--- app/views/home/_user_intention_survey_setup.html.haml
@@ -1,36 +1,0 @@
-- case $USER_INTENTION_SURVEY
-- when :accomplish
-  :javascript
-    jQuery(document).ready(function(){
-      var surveyName = #{$USER_INTENTION_SURVEY.to_json};
-      var options = {
-        surveyName: surveyName
-      };
-      new UserIntentionSurvey(options).bindTracker();
-    });
-- when :external_july_2014
-  :javascript
-    jQuery(document).ready(function(){
-      var surveyName = #{$USER_INTENTION_SURVEY.to_json};
-      var options = {
-        surveyName: surveyName,
-        remindUser: true,
-        tempDisableLength: 1,
-        forwardOnSubmit: true,
-        forwardToUrl: "https://www.surveymonkey.com/s/ZBX6T2Q",
-        submitForAnalytics: false
-      };
-      new UserIntentionSurvey(options).bindTracker();
-    });
-
-  :css
-    .user_intention_survey_text {
-      border: solid thick #338D0C !important;
-      padding: 15px;
-      border-radius: 5px;
-      background: lightGrey;
-    }
-    .user_intention_survey {
-      width: 450px !important;
-      -webkit-margin-end: 20px !important;
-    }
