--- app/views/home/account.html.erb
@@ -1,3 +1,0 @@
-<% @title = "Account Information" %>
-
-<%= render :partial => "users/details" %>
