--- app/views/home/user.html.erb
@@ -1,66 +1,0 @@
-<div class="leftHome">
-<div class="feeds">
-	<div class="header">Subscribed Feeds <img src="/images/feed-icon.png" width=25 style="float:right;position:relative;top:-20px;"></div>
-
-      					<ul>
-						<li><span class="feedLabel">Comment:</span> <a href="#">Xenopus anatomy and development => Cell</a><br>
-							<span class="quote">"I think the definition for this class would..."</span>
-						</li>
-						<li><span class="feedLabel">Change:</span> <a href="#">Pathway Ontology => Diabetes Pathway:Definition</a><br>
-							<span class="quote">"A broad condition with various manifestations that is associated..."</span>
-						</li>
-						<li><span class="feedLabel">Proposal:</span> <a href="#">NCI Thesaurus => Cell:Definition</a></li>
-						<li><span class="feedLabel">Addition:</span> <a href="#">NCI Thesaurus => Blood</a></li>
-						
-					</ul>
-		
-</div>
-<div class="stats">
-	<div class="header">Statistics</div>
-<ul>
-	<li>Total Number of Ontologies   	  	72</li>
-	<li>NCBO Library 		59</li>	
-	<li>Remote 		13</li>
-	<li>Number of Classes/Types 		300109*</li>
-</ul>
-*ontologies which have been parsed and indexed
-
-	
-</div>
-</div>
-
-
-<div class="myontologies">
-	<div class="header">My Ontologies <img src="/images/feed-icon.png" width=25 style="float:right;position:relative;top:-20px;"></div>
-                	<h4><a href="/Xenopus-anatomy-and-development">Xenopus anatomy and development =></a></h4>
-                	
-					<ul>
-						<li>7 new comments added</li>
-						<li>3 proposals added</li>
-					</ul>
-
-                		<h4><a href="/Sequence-types-and-features">Sequence types and features =></a></h4>
-                	
-					<ul>
-						<li>7 new comments added</li>
-						<li>3 proposals added</li>
-					</ul>
-
-                	<h4><a href="/Pathway-ontology">Pathway ontology =></a></h4>
-                	
-					<ul>
-						<li>7 new comments added</li>
-						<li>3 proposals added</li>
-					</ul>
-					
-					<h4>Proposals</h4>
-							<ul>
-						<li>Your proposal for <a href="/Xenopus-anatomy-and-development/XAO%3A1000009">Xenopus anatomy and development => adult</a> is still pending.</li>
-						<li>Your proposal for <a href="/Xenopus-anatomy-and-development/XAO%3A0000113">Xenopus anatomy and development => embryo</a> has been approved.</li>
-						<li>A proposal has been created for <a href="/Sequence-types-and-features/SO:1000132">Sequence types and features => mutation</a>.</li>						
-					</ul>
-		
-		
-	
-
-</div>
