--- app/views/notifier/endlessloop.html.erb
@@ -1,3 +1,0 @@
-Endless Loop at node
-
-<%=@node%>
