--- app/views/ontologies/_private_ontology.html.erb
@@ -1,10 +1,0 @@
-<div style="padding: 1em; width: 500px">
-  <h2>This ontology is private</h2>
-  <% ontology_authors = @ontology.contactName.nil? || @ontology.contactName.empty? ? "the authors" : @ontology.contactName%>
-  <p>
-    This ontology is private and only those who have been granted access by the ontology author may browse its classes.
-    Please contact <%= ontology_authors%> at <a href="mailto:<%= @ontology.contactEmail%>"><%= @ontology.contactEmail%></a> if you would like access.
-
-    <%if session[:user].nil?%>If you have already been granted access then please <a href="/login?redirect=<%= request.url%>">log in</a> and you will be taken to the class browser.<%end%>
-  </p>
-</div>
