--- app/views/search/_concept_preview.html.erb
@@ -1,17 +1,0 @@
-<fieldset style="width:300px;">
-<legend>Preview of <%=link_to "#{@concept.label_html},#{@concept.id}",uri_url(:ontology=>@ontology,:conceptid=>@concept.id),:target=>"_new"%></legend>
-<!--<div style="overflow:auto;width:300px"></div>-->
-<br>
-<table class="minimal">
-	<tbody>		
-<%@concept.properties.each_pair{|key,value| %>
-
-	<tr>
-	<td><%="#{remove_owl_notation(key)}:"%></td><td><%="#{value}"%></td>	
-	</tr>
-	
-<% } %>
-</tbody>
-</table>
-
-</fieldset>
