--- app/views/search/_concepts.html.erb
@@ -1,22 +1,0 @@
-<%if @concepts.size > 0%>
-	<div>
-		<div style="float:left;margin:10px;"id="select">
-			<select name="mapping[destination_id]" style="width:200px" size="18" onchange="$('#map_info').load('/search/concept_preview?ontology=<%=@ontology.id%>&id='+this.value);">
-				<%for option in @concepts%>
-				  <option value="<%=option[:conceptIdShort]%>"><%=option[:preferredName]%></option>
-				<%end%>
-			</select>
-		</div>
-		
-		<div style="float:left;margin:10px;"id="map_info"></div>
-		
-		<div style="margin: 10px; clear: both;">Comment<br/><%=text_area(:mapping, :comment, :size => "30x10")%></div>
-	</div>
-	
-	<div style="clear:both;white-space:nowrap">
-	  <%= select("mapping", "directionality", { "Bidirectional" => "bidirectional", "Unidirectional" => "unidirectional" })%><br><br>
-    <%= submit_tag "Create", :class=>'blueButton'%> &nbsp;&nbsp;
-	</div>
-<%else%>
-  <h1>No Results Found</h1>
-<%end%>
