--- app/views/search/search.html.erb
@@ -1,6 +1,0 @@
-<ul>
-<%for result in @results%>
-	<%=result.name%>
-
-<%end%>
-</ul>
