--- app/views/shared/_not_browsable.html.erb
@@ -1,21 +1,0 @@
-<script type="text/javascript">
-	jQuery(document).ready(function(){
-		var url = "/#TB_inline?height=400&width=600&inlineId=not_browsable_content&modal=true"
-        tb_show("", url);
-	});
-</script>
-<div style="min-height: 600px">
-<div id="not_browsable_content" style="display:none;">
-    <h1 style="font-size: 200%; font-weight: bold; padding-bottom: .8em;">Not Browsable</h1>
-	<p><span style="font-size: large; line-height: 1.2em;">
-		<%= ErrorHelpers::not_browsable_text(@concept) %><br><br>
-		From here you can:<br>
-		<span style="line-height: 1.5em;">
-		  <ul>
-			<li style="list-style-position: inside; list-style-type: disc; margin-left: 1em;"><a href="#" onclick="window.history.back(); return false;">Go back</a> to your previous page</li>
-			<li style="list-style-position: inside; list-style-type: disc; margin-left: 1em;">View <%=link_to @ontology.displayLabel, visualize_url(:ontology => @ontology.id)%>, which contains <strong><%=@concept.name%></strong></li>
-		  </ul>
-		</span>
-	</span></p>
-</div>
-</div>
